build: 优化 translate markdown 里的 html 节点 translate
