+++ package.json
@@ -48,14 +48,9 @@
     "lint-staged": "^13.0.3",
     "mdast": "^3.0.0",
     "npm-check-updates": "^16.3.3",
-    "rehype-katex": "^6.0.2",
-    "rehype-parse": "^8.0.4",
-    "rehype-remark": "^9.1.2",
-    "rehype-stringify": "^9.0.3",
     "remark-gfm": "^3.0.1",
     "remark-math": "^5.1.1",
     "remark-parse": "^10.0.1",
-    "remark-rehype": "^10.1.0",
     "remark-stringify": "^10.0.2",
     "rimraf": "^3.0.2",
     "turbo": "^1.5.4",
