+++ scripts/translate-md/index.ts
@@ -1,3 +1,3 @@
 export * from './translate-md'
 export * from './translate-html'
-export * from './translate-utils'
+export * from './utils'
