+++ scripts/translate-md/translate-html.ts
@@ -1,5 +1,6 @@
-import type { GoogleTranslateOptions } from './translate-utils'
-import { betterTranslate } from './translate-utils'
+import { createReplacer } from './replacer'
+import type { GoogleTranslateOptions } from './utils'
+import { betterTranslate } from './utils'
 
 /**
  * 翻译 html 内容配置
@@ -19,46 +20,20 @@ interface TranslateHtmlOptions extends GoogleTranslateOptions {
 export async function translateHtml(options: TranslateHtmlOptions): Promise<string> {
   const { htmlContent, sourceLang, targetLang, useSystemProxy } = options
 
-  const { unified } = await import('unified')
-  const { default: parse } = await import('rehype-parse')
-  const { default: stringify } = await import('rehype-stringify')
-  const { visit } = await import('unist-util-visit')
-
-  const processor = unified()
-    .use(parse, {
-      fragment: true,
-      verbose: true,
-    })
-    .use(stringify, {
-      allowDangerousHtml: true,
-    })
-
-  const tree = processor.parse(htmlContent)
-
-  // html 里的纯文字文本
-  const translateTexts: string[] = []
-
-  // 收集纯文字文本
-  visit(tree, 'text', (node) => {
-    translateTexts.push(node.value)
-  })
-
-  // 翻译纯文字文本
-  const translateTextsResult: string[] = await Promise.all(translateTexts.map((text) => {
-    return betterTranslate(text, {
+  const translate = (str: string) => {
+    return betterTranslate(str, {
       sourceLang,
       targetLang,
       useSystemProxy,
     })
-  }))
+  }
 
-  // 替换纯文字文本
-  let index = 0
-  visit(tree, 'text', (node) => {
-    node.value = translateTextsResult[index]
-    index++
-  })
+  const htmlTextRegExp = /(<[\w\W]+?>\s*)([\w\W]*?)(?=\s*<[\w\W]+?>)/g
+  const htmlTranslateFn = createReplacer(htmlContent, htmlTextRegExp, (matchResult, translateResult) => {
+    const forTranslateContent = matchResult[2] || ''
+    const forReturnReplacerContent = `${matchResult[1]}${translateResult}`
+    return [forTranslateContent, forReturnReplacerContent]
+  }, translate)
 
-  return processor.stringify(tree)
+  return htmlTranslateFn()
 }
-
